pyproject.toml
[project]
name = "diagnostics-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "docx2txt>=0.9",
    "fastapi>=0.116.1",
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-nvidia-ai-endpoints>=0.3.13",
    "langchain-openai>=0.3.28",
    "langchain-postgres>=0.0.15",
    "langchain-qdrant>=0.2.0",
    "langfuse>=3.2.1",
    "psycopg[binary]>=3.2.9",
    "pydantic-ai>=0.1.0",
    "pydantic-ai-slim[duckduckgo]>=0.4.9",
    "pypdf>=5.9.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.0",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
]
